fix: different files for types
